refactor(ingestion-consumer): replace the committer trait with a commit pacer
`Committer` and `CommitManager` named something that never commits: it
holds the latest frontier per partition and hands them out once per
interval. The struct is now `CommitPacer` and the interval-gated method
is `take_due`, which says what leaves and why.

The trait had one implementation and a null stand-in for the sentinel's
tests. The sentinel and the commit monitor now hold the pacer directly,
and the sentinel's tests wrap a real one.
