+++ rust/ingestion-consumer/src/commit_monitor.rs
@@ -13,7 +13,6 @@ use rdkafka::consumer::{Consumer, StreamConsumer};
 use tokio::task::JoinHandle;
 use tracing::warn;
 
-use crate::commit_manager::Committer;
 use crate::commit_sentinel::CommitSentinel;
 use crate::order_sentinel::SentinelContext;
 
@@ -31,9 +30,9 @@ impl Drop for AbortOnDrop {
 
 /// Run the monitor until shutdown. Aborted when the guard drops, so a
 /// consumer torn down mid-test doesn't keep the rdkafka client alive.
-pub fn spawn_commit_monitor<C: Committer + 'static>(
+pub fn spawn_commit_monitor(
     consumer: Arc<StreamConsumer<SentinelContext>>,
-    sentinel: Arc<CommitSentinel<C>>,
+    sentinel: Arc<CommitSentinel>,
     handle: Handle,
 ) -> AbortOnDrop {
     AbortOnDrop(tokio::spawn(run_commit_monitor(consumer, sentinel, handle)))
@@ -43,9 +42,9 @@ pub fn spawn_commit_monitor<C: Committer + 'static>(
 /// assignment (an OffsetFetch round trip) and feed them to the commit
 /// sentinel, which compares them against attempted commits and stamps the
 /// last-successful-commit gauge on progress.
-async fn run_commit_monitor<C: Committer>(
+async fn run_commit_monitor(
     consumer: Arc<StreamConsumer<SentinelContext>>,
-    sentinel: Arc<CommitSentinel<C>>,
+    sentinel: Arc<CommitSentinel>,
     handle: Handle,
 ) {
     loop {
