+++ rust/ingestion-consumer/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod aperture;
 pub mod batcher;
-pub mod commit_manager;
 pub mod commit_monitor;
+pub mod commit_pacer;
 pub mod commit_sentinel;
 pub mod config;
 pub mod consumer;
